refactor: rename reservations pages
