+++ apps/admin-ui/src/spa/reservations/[id]/eventStyleGetter.ts
