+++ apps/admin-ui/src/spa/reservations/[id]/hooks/index.ts
@@ -4,7 +4,6 @@ import {
   ReservationTypeChoice,
   useReservationQuery,
   useReservationsByReservationUnitQuery,
-  useRecurringReservationQuery,
   ReservationUnitNode,
   ReservationNode,
 } from "@gql/gql-types";
@@ -13,6 +12,7 @@ import { toApiDate } from "common/src/common/util";
 import { errorToast } from "common/src/common/toast";
 import { base64encode, filterNonNullable } from "common/src/helpers";
 import { type CalendarEventType } from "../Calendar";
+import { useRecurringReservations } from "@/hooks";
 
 const getEventName = (
   eventType?: ReservationTypeChoice | null,
@@ -127,34 +127,6 @@ export function useReservationData(
       .map((r) => convertReservationToCalendarEvent(r, blockedName)) ?? [];
 
   return { ...rest, events };
-}
-
-/// @param recurringPk fetch reservations related to this pk
-/// @param state optionally only fetch some reservation states
-/// @param limit allows to over fetch: 100 is the limit per query, larger amounts are done with multiple fetches
-export function useRecurringReservations(recurringPk?: number) {
-  const { t } = useTranslation();
-
-  const id = base64encode(`RecurringReservationNode:${recurringPk}`);
-  const { data, loading, refetch } = useRecurringReservationQuery({
-    skip: !recurringPk,
-    fetchPolicy: "cache-and-network",
-    nextFetchPolicy: "cache-first",
-    variables: { id },
-    onError: () => {
-      errorToast({ text: t("errors.errorFetchingData") });
-    },
-  });
-
-  const { recurringReservation } = data ?? {};
-  const reservations = filterNonNullable(recurringReservation?.reservations);
-
-  return {
-    loading,
-    reservations,
-    recurringReservation,
-    refetch,
-  };
 }
 
 /// @param id fetch reservation related to this pk
