+++ apps/admin-ui/src/spa/reservations/router.tsx
@@ -1,16 +1,17 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import RequestedReservations from "./RequestedReservations";
-import AllReservations from "./AllReservations";
-import RequestedReservation from "./requested/RequestedReservation";
-import EditPage from "./EditPage";
+import { RequestedPage } from "./requested";
+import { ListReservationsPage } from ".";
+import { EditPage } from "./[id]/edit";
+import { ReservationPage } from "./[id]";
 
 // TODO there is no index? (all and requested works like index but not really)
 const ReservationsRouter = (): JSX.Element => (
   <Routes>
-    <Route path="requested" element={<RequestedReservations />} />
-    <Route path="all" element={<AllReservations />} />
-    <Route path=":id" element={<RequestedReservation />} />
+    <Route index element={<ListReservationsPage />} />
+    <Route path="requested" element={<RequestedPage />} />
+    <Route path="all" element={<ListReservationsPage />} />
+    <Route path=":id" element={<ReservationPage />} />
     <Route path=":id/edit" element={<EditPage />} />
   </Routes>
 );
